% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_rivers}
\alias{cb_rivers}
\title{Cape Breton Rivers}
\format{
A \code{sf} object with 8 fields and 2268 features.
\describe{
\item{feat_code}{The feature code}
\item{feat_desc}{The feature description}
\item{rivname_1}{The first river name}
\item{rivname_2}{The second river name}
\item{zvalue}{The z value - (depth? in m)}
\item{shape_len}{The length of the shape in km}
\item{shape_fid}{The shape feature ID}
\item{geometry}{The geometry of the shape}
}
}
\usage{
cb_rivers
}
\description{
A \code{sf} object that contains all the rivers in Cape Breton, Nova Scoatia.
The CRS is \code{WGS84(DD)}
}
\keyword{datasets}
