% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cheticamp}
\alias{cheticamp}
\title{Cheticamp}
\format{
A \code{sf} object with 0 fields and 1 features.
\describe{
\item{geometry}{The geometry of the shape}
}
}
\usage{
cheticamp
}
\description{
A \code{sf} object that contains the Cheticamp habour and the associated lakes and rivers
in the area.
The CRS is \verb{NAD83(CSRS) / UTM zone 20N + CGVD28 height}
}
\keyword{datasets}
