% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eastern_canada}
\alias{eastern_canada}
\title{Eastern Canada}
\format{
A \code{sf} object with 0 fields and 1 features.
\describe{
\item{pruid}{The shape feature ID}
\item{prname}{The province names both English and French}
\item{prename}{The province names in Enlgish}
\item{prfname}{The province names in French}
\item{preabbr}{The province abbrivations in Enlgish}
\item{prfabbr}{The province abbrivations in French}
\item{geometry}{The geometry of the shape}
}
}
\usage{
eastern_canada
}
\description{
A \code{sf} object that contains eastern Canada including Quebec, New Brunswick, Nova Scotia,
Prince Edward Island, and Newfoundland and Labrador.
The CRS is \code{PCS_Lambert_Conformal_Conic}
}
\keyword{datasets}
