% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fathom.R
\name{fathom}
\alias{fathom}
\title{function to quickly format files from Innovasea fathom offloads}
\usage{
fathom(x, m, n = 12)
}
\arguments{
\item{x}{Path containing Fathom csv files.}

\item{m}{Google sheet metadata}

\item{n}{Number of header rows to skip when reading files.}
}
\description{
function to quickly format files from Innovasea fathom offloads
}
