% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab}
\alias{lab}
\title{Lab}
\format{
A \code{RasterLayer}
}
\usage{
lab
}
\description{
A \code{RasterLayer} object that contains ????
The CRS is \code{+proj=longlat +datum=WGS84 +no_defs}
}
\keyword{datasets}
