% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mil}
\alias{mil}
\title{Milipsigate Lake}
\format{
A \code{sf} object with 0 fields and 1 features.
\describe{
\item{Name}{The feature name}
\item{Description}{The feature description}
\item{geometry}{The geometry of the shape}
}
}
\usage{
mil
}
\description{
A \code{sf} object that contains the Milipsigate Lake, Nova Scotia.
}
\details{
The CRS is \verb{WGS 84}
}
\keyword{datasets}
