% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nshires}
\alias{nshires}
\title{Nova Scotia - Broad}
\format{
A \code{sf} object with 0 fields and 1 features.
\describe{
\item{geometry}{The geometry of the shape}
}
}
\usage{
nshires
}
\description{
A \code{sf} object that contains Nova Scotia in broad detail.
}
\details{
The CRS is \verb{WGS 84}
}
\keyword{datasets}
