% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otnr.R
\name{otnr}
\alias{otnr}
\title{function to quickly format files from OTN data extracts}
\usage{
otnr(parquet)
}
\arguments{
\item{parquet}{is the parquet file}
}
\description{
function to quickly format files from OTN data extracts
}
