% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Study Rivers of Interest}
\format{
A \code{sf} object with 1 fields and 764 features.
\describe{
\item{river}{The name of the river}
\item{geometry}{The geometry of the shape}
}
}
\usage{
rivers
}
\description{
A \code{sf} object that contains Nova Scotia study rivers of interest.
}
\details{
The CRS is \verb{NAD83(CSRS) / UTM zone 20N}
}
\keyword{datasets}
