% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wallace}
\alias{wallace}
\title{Wallace River}
\format{
A \code{sf} object with 7 fields and 600 features.
\describe{
\item{feat_code}{The feature code}
\item{feat_desc}{The feature description}
\item{rivname_1}{The first river name}
\item{rivname_2}{The second river name}
\item{zvalue}{The z value - (depth? in m)}
\item{shape_len}{The length of the shape in km}
\item{shape_fid}{The shape feature ID}
\item{geometry}{The geometry of the shape}
}
}
\usage{
wallace
}
\description{
A \code{sf} object that contains all the Wallace River, Nova Scotia.
The CRS is \code{WGS84(DD)}
}
\keyword{datasets}
