% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fathom.R
\name{fathom}
\alias{fathom}
\title{function to quickly format files from Innovasea fathom offloads}
\usage{
fathom(x, n = 12)
}
\arguments{
\item{x}{is the fathom offload file}
}
\description{
function to quickly format files from Innovasea fathom offloads
}
